requirements.txt
pydantic>=2.0.0
PyYAML>=6.0
Jinja2>=3.0.0
ruff>=0.1.5
requests>=2.25.0
aiohttp>=3.8.0
httpx>=0.24.0
